Create utils module
So that file saving is independent of QT
